requirements.txt
pymongo
python-dotenv
hashlib
uuid
datetime
